Various code cleanups
