+++ app/src/main/java/moe/apex/rule34/util/Ui.kt
@@ -1,7 +1,14 @@
 package moe.apex.rule34.util
 
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.animation.slideInVertically
+import androidx.compose.animation.slideOutVertically
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
@@ -9,7 +16,13 @@ import androidx.compose.material3.LargeTopAppBar
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBarScrollBehavior
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.MutableIntState
+import androidx.compose.runtime.MutableState
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
 import androidx.navigation.NavController
+import moe.apex.rule34.image.Image
+import moe.apex.rule34.largeimageview.LargeImageView
 
 
 @Composable
@@ -33,4 +46,38 @@ fun TitleBar(
             }
         }
     )
+}
+
+
+@Composable
+fun FullscreenLoadingSpinner() {
+    Row(
+        Modifier.fillMaxSize(),
+        verticalAlignment = Alignment.CenterVertically,
+        horizontalArrangement = Arrangement.Center
+    ) {
+        CircularProgressIndicator()
+    }
+}
+
+
+@Composable
+fun AnimatedVisibilityLargeImageView(
+    shouldShowLargeImage: MutableState<Boolean>,
+    navController: NavController,
+    initialPage: MutableIntState,
+    allImages: List<Image>
+) {
+    AnimatedVisibility(
+        visible = shouldShowLargeImage.value,
+        enter = slideInVertically(initialOffsetY = { it }),
+        exit = slideOutVertically(targetOffsetY = { it })
+    ) {
+        LargeImageView(
+            navController,
+            initialPage,
+            shouldShowLargeImage,
+            allImages
+        )
+    }
 }
